Realized offers on map, hurray!
