+++ js/main.js
@@ -1,5 +1,4 @@
 import './modules/data.js';
-// import './modules/offer.js';
 import './modules/form.js';
 import './modules/validation.js';
 import './modules/map.js';
